fix: Keep removing py.py file when initializing may lead to endless run time.

+++ hml/generators.py
@@ -87,8 +87,8 @@ class Madgraph5:
                 stderr=subprocess.PIPE,
             )
 
-            # Remove py.py file
-            while process.poll() is None:
+            # Remove py.py file after initialization
+            if process.poll() is not None:
                 if Path("py.py").exists():
                     Path("py.py").unlink()
